Get sizes of data types

+++ 0x00-hello_world/6-size.c
@@ -6,11 +6,11 @@
 */
 int main(void)
 {
-printf("Size of a char: %ld byte(s)\n", sizeof(char));
-printf("Size of an int: %ld byte(s)\n", sizeof(int));
-printf("Size of a long int: %ld byte(s)\n", sizeof(long));
-printf("Size of a long long: %ld byte(s)\n", sizeof(long long));
-printf("Size of a float: %ld byte(s)\n", sizeof(float));
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+printf("Size of an int: %ld byte(s)\n", (unsigned long)sizeof(int));
+printf("Size of a long int: %ld byte(s)\n", (unsigned long)sizeof(long));
+printf("Size of a long long: %ld byte(s)\n", (unsigned long)sizeof(long long));
+printf("Size of a float: %ld byte(s)\n", (unsigned long)sizeof(float));
 return (0);
 }
